requirements.txt
Django==1.5
Pillow==2.2.2
South==0.8.4
Unipath==1.0
argparse==1.2.1
cmsplugin-filer==0.9.5
django-classy-tags==0.4
django-cms==2.4.3
django-filer==0.9.5
django-mptt==0.5.2
django-polymorphic==0.5.3
django-sekizai==0.7
easy-thumbnails==1.4
html5lib==0.999
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
